[new] adding independent files for methods and creating new branch stage
